+++ History.md
@@ -1,5 +1,10 @@
 # ratelimiter Change Log.
 
+## 3.0.0 / 2016-03-01
+* rewritten using es6 with babel
+* added eslint airbnb configuration
+* introduced adapters (redis, memory and null)
+
 ## 1.0.3 / 2014-06-06
 ==================